refactor(#1116): split the two time comparators out of their over-threshold files
types.rs (2920) and write_engine/mutation.rs (1854) are both far over the
800-line source target, and #3935's per-arm authority notes grew each of them,
which the file-size ratchet correctly FAILs. Split by responsibility rather
than opting out of the campsite rule:

  * types::value_ord   — `impl PartialOrd for Value` and its contract doc
    (two documented divergences from the derived PartialEq, plus the
    time/timestamp asymmetry). PRIVATE module: a trait impl applies crate-wide
    regardless of module visibility, so this adds no public surface.
  * write_engine::clustering_order — `compare_values`, the comparator BOTH
    routes to clustering order reach (ClusteringKey's Ord for the memtable
    BTreeMap, and the schema-aware compare the merge uses for output row
    order), hence the strict-total-order requirement documented there.

types.rs 2920 -> 2879, mutation.rs 1854 -> 1781; the two new files are 67 and
110 lines. No behaviour change: cqlite-core's 3711 lib tests pass unchanged.

Refs #1116 #3935
